refactor: TokenProvider 인터페이스에 의존하도록
